src/nyc_traffic_collision_analysis_and_prediction/app-server/requirements.txt
numpy
flask
pymongo
pyspark
flask-cors
pandas
huggingface_hub
joblib
python-dotenv
scikit-learn